複数の指板図を描画できるようにする
