+++ app/javascript/controllers/index.js
@@ -2,13 +2,10 @@
 // Run that command whenever you add a new controller or create them with
 // ./bin/rails generate stimulus controllerName
 
-import { application } from "./application"
+import { application } from "./application";
 
-import FretboardController from "./fretboard_controller"
-application.register("fretboard", FretboardController)
+import HelloController from "./hello_controller";
+application.register("hello", HelloController);
 
-import HelloController from "./hello_controller"
-application.register("hello", HelloController)
-
-import ScoreController from "./score_controller"
-application.register("score", ScoreController)
+import ScoreController from "./score_controller";
+application.register("score", ScoreController);
