+++ app/views/scores/_form.html.erb
@@ -1,19 +1,19 @@
-<div data-controller="fretboard">
+<div data-controller="score">
   <%= bootstrap_form_with(model: score) do |f| %>
     <%= f.text_field :title %>
-    <%= f.text_area :score_code, data: { fretboard_target: "output" } %>
+    <%= f.text_area :score_code, data: { score_target: "output" } %>
     <%= f.check_box :is_public %>
     <%= f.primary %>
   <% end %>
 
-  <%= bootstrap_form_tag do |f| %>
-    <%= f.select :start_fret, [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23 ], include_blank: "Select a Start Fret", floating: true, data: { fretboard_target: "start_fret" } %>
-    <%= f.select :end_fret, [ 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,24 ], include_blank: "Select a End Fret", floating: true, data: { fretboard_target: "end_fret" } %>
+  <%= bootstrap_form_with do |f| %>
+    <%= f.select :startFret, [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23 ], include_blank: "Select a Start Fret", floating: true, wrapper: { data: { score_target: "startFret" } } %>
+    <%= f.select :endFret, [ 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23,24 ], include_blank: "Select a End Fret", floating: true, wrapper: { data: { score_target: "endFret" } } %>
   <% end %>
 
-  <button data-action="click->fretboard#generateFretboardCode">fetch</button>
-  <button data-action="click->fretboard#fetchCode">fetch</button>
+  <button data-action="click->score#generateStage">fetch</button>
 
-  <button data-action="click->fretboard#add">Add</button>
-  <div data-fretboard-target="output" id="scoreContainer"></div>
+  <button data-action="click->score#AddFretboard">Add</button>
+  <div data-score-target="output" id="scoreContainer"></div>
+
 </div>
